+++ prov/efa/Makefile.include
@@ -50,7 +50,8 @@ _efa_files = \
 	prov/efa/src/rxr/rxr_rma.c	\
 	prov/efa/src/rxr/rxr_msg.c	\
 	prov/efa/src/rxr/rxr_pkt_entry.c \
-	prov/efa/src/rxr/rxr_pkt_type_rts.c
+	prov/efa/src/rxr/rxr_pkt_type_rts.c \
+	prov/efa/src/rxr/rxr_pkt_type_data.c
 
 _efa_headers = \
 	prov/efa/src/efa.h \
